Reorganise the Logging directory
